requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
anthropic
markdown
beautifulsoup4
sqlalchemy
# psycopg2-binary
openai
scikit-learn